fix(clean-up): refactoring and cleaning up application
