+++ utils/supabase/hypercerts.ts
@@ -98,55 +98,3 @@ export type GetHypercertsByIdResponse = {
 	data: HypercertData | null;
 	error: PostgresError | null;
 };
-
-// TODO: Delete this as we migrated to GraphQL
-export const fetchHypercertById = async (
-	hypercert_id: string,
-): Promise<GetHypercertsByIdResponse> => {
-	const supabase = createClient();
-
-	if (!supabase) {
-		throw new Error("Supabase client is not initialized");
-	}
-
-	try {
-		const { data, error } = await supabase
-			.from("claims")
-			.select(`
-				*,
-				metadata (
-					allow_list_uri,
-					contributors,
-					description,
-					external_url,
-					id,
-					image,
-					impact_scope,
-					impact_timeframe_from,
-					impact_timeframe_to,
-					name,
-					parsed,
-					properties,
-					rights,
-					uri,
-					work_scope,
-					work_timeframe_from,
-					work_timeframe_to
-				),
-				fractions!inner (id, units, owner_address)
-			`)
-			.eq("hypercert_id", hypercert_id);
-
-		const hypercertData = data?.[0];
-
-		return {
-			data: hypercertData as HypercertData,
-			error: error as PostgresError | null,
-		};
-	} catch (error) {
-		return {
-			data: null,
-			error: error as PostgresError,
-		};
-	}
-};
